plotting: renaming, included background when plotting lens modelling
